update nested directory untuk generate controller , libraries, dan view

+++ system/filesystem/FileSystem.php
@@ -51,55 +51,99 @@ class FileSystem
 
     public function create_controller($name = '')
     {
-        // problem belum bisa buat nested directory
-        $new_filename = $name.".php";
+        // config structure directory nested
+        $structure = explode("/", $name);
+        $file = end($structure);
+        $nama_file = $file; # nama file
 
+        // remove last index of array
+        if (count($structure) >= 1) {
+            array_pop($structure);
+        }
+        
+        $new_filename = $file.".php";
+
+        // get new structur
+        $new_structur = implode("/", $structure);
+
+        $path_core = dirname(__DIR__, 2)  . "/apps/controllers/";
         $path_default_file = dirname(__DIR__, 1) . "/defaults/controller/BasicController.php";
-        $path_new_file = dirname(__DIR__, 2)  . "/apps/controllers/". $new_filename;
+        $path_new_file = dirname(__DIR__, 2)  . "/apps/controllers/". $new_structur . DIRECTORY_SEPARATOR . $new_filename;
 
         if (file_exists($path_default_file)) {
             if (file_exists($path_new_file)) {
                 return "file " . $path_new_file ." controller sudah ada";
             }else{
+               
+                // check jika path folder dan file tidak ada;
+                while (!file_exists($path_core . $new_structur)){
+                    mkdir( $path_core . $new_structur, 777, true);
+                    break;
+                }
+                
+                // check buat nested directory jika tidak ada
+                if(file_exists($path_core . $new_structur)){
+                    $default_name = "BasicController";
+    
+                    $file = file_get_contents($path_default_file);          # read file
+                    $file = str_replace($default_name, $nama_file , $file);      # replace string
+    
+                    file_put_contents($path_new_file , $file);              # create file
+    
+                    return "buat file " . $path_new_file . " controller  berhasil";
+                }
 
-                $default_name = "BasicController";
-
-                $file = file_get_contents($path_default_file);          # read file
-                $file = str_replace($default_name, $name , $file);      # replace string
-
-                file_put_contents($path_new_file , $file);              # create file
-
-                return "buat file " . $path_new_file . " controller  berhasil";
             }
         }
     }
 
     public function create_libraries($name = '')
     {
-        // problem belum bisa buat nested directory
-        $new_filename = $name.".php";
+        // config structure directory nested
+        $structure = explode("/", $name);
+        $file = end($structure);
+        $nama_file = $file; # nama file
 
+        // remove last index of array
+         if (count($structure) >= 1) {
+            array_pop($structure);
+        }
+        $new_filename = $file.".php";
+
+        // get new structur
+        $new_structur = implode("/", $structure);
+
+        $path_core = dirname(__DIR__, 2)  . "/apps/libraries/";
         $path_default_file = dirname(__DIR__, 1) . "/defaults/libraries/BasicLibraries.php";
-        $path_new_file = dirname(__DIR__, 2)  . "/apps/libraries/".$new_filename;
+        $path_new_file = dirname(__DIR__, 2)  . "/apps/libraries/". $new_structur . DIRECTORY_SEPARATOR . $new_filename;
 
         if (file_exists($path_default_file)) {
             if (file_exists($path_new_file)) {
                 return "file " . $path_new_file ." libraries sudah ada";
             }else{
-                $default_name = "BasicLibraries";
+                // check jika path folder dan file tidak ada;
+                while (!file_exists($path_core . $new_structur)){
+                    mkdir( $path_core . $new_structur, 777, true);
+                    break;
+                }
 
-                $file = file_get_contents($path_default_file);          # read file
-                $file = str_replace($default_name, $name , $file);      # replace string
+                 // check buat nested directory jika tidak ada
+                 if(file_exists($path_core . $new_structur)){
+                    $default_name = "BasicLibraries";
 
-                file_put_contents($path_new_file , $file);              # create file
-                return "buat file " . $path_new_file . " libraries  berhasil dibuat";
+                    $file = file_get_contents($path_default_file);              # read file
+                    $file = str_replace($default_name, $nama_file , $file);     # replace string
+
+                    file_put_contents($path_new_file , $file);                  # create file
+                    return "buat file " . $path_new_file . " libraries  berhasil dibuat";
+                }
             }
         }
     }
 
     public function create_models($name = '')
     {
-        // problem belum bisa buat nested directory
+        // khusus models tidak bisa nested
         $new_filename = $name.".php";
 
         $path_default_file = dirname(__DIR__, 1) . "/defaults/models/BasicModels.php";
@@ -147,23 +191,44 @@ class FileSystem
 
     public function create_view($name = '')
     {
-        // problem belum bisa buat nested directory
-        $new_filename = $name.".php";
+        // config structure directory nested
+        $structure = explode("/", $name);
+        $file = end($structure);
+        $nama_file = $file; # nama file
 
+        // remove last index of array
+         if (count($structure) >= 1) {
+            array_pop($structure);
+        }
+        $new_filename = $file.".php";
+
+        // get new structur
+        $new_structur = implode("/", $structure);
+
+        $path_core = dirname(__DIR__, 2)  . "/apps/views/";
         $path_default_file = dirname(__DIR__, 1) . "/defaults/views/BasicView.php";
-        $path_new_file = dirname(__DIR__, 2)  . "/apps/views/".$new_filename;
+        $path_new_file = dirname(__DIR__, 2)  . "/apps/views/". $new_structur . DIRECTORY_SEPARATOR . $new_filename;
 
         if (file_exists($path_default_file)) {
             if (file_exists($path_new_file)) {
                 return "file " . $path_new_file ." view sudah ada";
             }else{
-                $default_name = "Basic View";
+                 // check jika path folder dan file tidak ada;
+                 while (!file_exists($path_core . $new_structur)){
+                    mkdir( $path_core . $new_structur, 777, true);
+                    break;
+                }
 
-                $file = file_get_contents($path_default_file);          # read file
-                $file = str_replace($default_name, $name , $file);      # replace string
+                // check buat nested directory jika tidak ada
+                if(file_exists($path_core . $new_structur)){
+                    $default_name = "Basic View";
 
-                file_put_contents($path_new_file , $file);              # create file
-                return "buat file " . $path_new_file . " views berhasil";
+                    $file = file_get_contents($path_default_file);              # read file
+                    $file = str_replace($default_name, $nama_file , $file);     # replace string
+
+                    file_put_contents($path_new_file , $file);                  # create file
+                    return "buat file " . $path_new_file . " views berhasil";
+                }
             }
             
         }
